+++ R/st_nn_pnt_geo.R
@@ -6,6 +6,10 @@
   maxdist = units::set_units(maxdist, "m")
   maxdist = as.numeric(maxdist)
 
+  # result = getIdsDists(x_coord, y_coord, k, maxdist)
+  # ids = result[[1]]
+  # dist_matrix = result[[2]]
+
   ids = matrix(NA, nrow = x_features, ncol = k)
   dist_matrix = matrix(NA, nrow = x_features, ncol = k)
 
@@ -14,26 +18,20 @@
 
   for(i in 1:x_features) {
 
-    # dists = sf::st_distance(x[i, ], y)[1, ]
-    dists = mapply(
-      getDistance,
-        x_coord[i, 1],
-        x_coord[i, 2],
-        y_coord[, 1],
-        y_coord[, 2]
-      )
+    dists = getDistancesN(x_coord[i, ], y_coord)
     ids1 = order(dists)[1:k]
-    dists1 = dists[ids1]
-    # dists1 = units::set_units(dists1, "m")
-    ids1[dists1 > maxdist] = NA
-    dists1[dists1 > maxdist] = NA
     ids[i, ] = ids1
-    dist_matrix[i, ] = dists1
+    dist_matrix[i, ] = dists[ids1]
 
     # Progress
     utils::setTxtProgressBar(pb, i)
 
   }
+
+  ids[dist_matrix > maxdist] = NA
+  dist_matrix[is.na(ids)] = NA
+
+  cat("\n")
 
   return(list(ids, dist_matrix))