+++ NEWS.md
@@ -5,7 +5,7 @@
 # nngeo 0.1.5
 
 * Added tests for 'st_nn'
-* Added progress bar for 'st_nn' and 'st_connect'
-* Added C code (Vincenty) for 'st_nn' & lon-lat points
+* Added progress bar for 'st_nn' (except when using projected points method) and for 'st_connect'
+* Added C code (Vincenty) for 'st_nn' lon-lat points method
 * Fixed sample size mistake in 'st_nn' example
 * Fixed error when using 'sfc' objects